Flatten media attachments into Hunter
