+++ lib/hunter/attachment.ex
@@ -2,8 +2,7 @@ defmodule Hunter.Attachment do
   @moduledoc """
   Attachment entity
 
-  This module defines a `Hunter.Attachment` struct and the main functions
-  for working with Attachments.
+  This module defines a `Hunter.Attachment` struct.
 
   ## Fields
 
@@ -26,7 +25,6 @@ defmodule Hunter.Attachment do
   available and local `url` is missing
 
   """
-  alias Hunter.Api.HTTPClient
 
   @type t :: %__MODULE__{
           id: String.t(),
@@ -54,79 +52,4 @@ defmodule Hunter.Attachment do
     :description,
     :blurhash
   ]
-
-  @doc """
-  Upload a media attachment
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `file` - media to be uploaded
-    * `options` - option list
-
-  ## Options
-    * `description` - plain-text description of the media for accessibility (max 420 chars)
-    * `focus` - two floating points, comma-delimited.
-
-  **Note:** the v2 media endpoint processes large files asynchronously: the
-  returned attachment's `url` may be `nil` until the server finishes
-  processing (HTTP 202). The `id` can be attached to a status with
-  `create_status` as soon as processing completes.
-
-  """
-  @spec upload_media(Hunter.Client.t(), Path.t(), Keyword.t()) :: Hunter.Attachment.t()
-  def upload_media(conn, file, options \\ []) do
-    HTTPClient.upload_media(conn, file, options)
-  end
-
-  @doc """
-  Retrieve a media attachment, to check the processing status of an
-  asynchronous upload
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - attachment identifier
-
-  """
-  @spec media_attachment(Hunter.Client.t(), String.t() | non_neg_integer) :: Hunter.Attachment.t()
-  def media_attachment(conn, id) do
-    HTTPClient.media_attachment(conn, id)
-  end
-
-  @doc """
-  Update a media attachment, before it is attached to a status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - attachment identifier
-    * `options` - option list
-
-  ## Options
-
-    * `description` - plain-text description of the media for accessibility
-    * `focus` - two floating points between -1.0 and 1.0, comma-delimited
-    * `thumbnail` - path of a custom thumbnail image
-
-  """
-  @spec update_media(Hunter.Client.t(), String.t() | non_neg_integer, Keyword.t()) ::
-          Hunter.Attachment.t()
-  def update_media(conn, id, options \\ []) do
-    HTTPClient.update_media(conn, id, options)
-  end
-
-  @doc """
-  Delete a media attachment that is not currently attached to a status
-
-  ## Parameters
-
-    * `conn` - connection credentials
-    * `id` - attachment identifier
-
-  """
-  @spec delete_media(Hunter.Client.t(), String.t() | non_neg_integer) :: boolean
-  def delete_media(conn, id) do
-    HTTPClient.delete_media(conn, id)
-  end
 end
